Fix some warnings, and move Menu, MenuItem, etc out of a sub-folder, not really necessary.
