+++ docs/custom.md
@@ -3,5 +3,6 @@
 
 The Custom control type covers a multitude of non-standard controls. Mostly these have been wrapped up in a single 
 type - Custom, but where specific implementations are known, then sub-classes of Custom have been created - for 
-example [Word](Word) has a AutomationDocumentPage, which is different enough from a standard Custom 
+example [Word](word.md) has a AutomationDocumentPage, which is different enough
+ from a standard Custom 
 control to have a specific implementation.
