+++ docs/swt.md
@@ -1,0 +1,1 @@
+Unfortunately it looks like `swt` is largely opaque to the automation APIs.
