+++ docs/word.md
@@ -1,4 +1,4 @@
-Word has a nice control hierarchy, and uses 'Pages' which are Custom controls.
+`Word` has a nice control hierarchy, and uses 'Pages' which are Custom controls.
 
 ![UISpy for Word](images/word.png)
