+++ src/main/java/mmarquee/automation/controls/AutomationControlFactory.java
@@ -19,9 +19,6 @@ import mmarquee.automation.AutomationException;
 import mmarquee.automation.ControlType;
 import mmarquee.automation.Element;
 import mmarquee.automation.ElementNotFoundException;
-import mmarquee.automation.controls.menu.MainMenu;
-import mmarquee.automation.controls.menu.Menu;
-import mmarquee.automation.controls.menu.MenuItem;
 import mmarquee.automation.pattern.PatternNotFoundException;
 
 /**
