+++ src/main/java/mmarquee/automation/controls/Menu.java
@@ -14,14 +14,12 @@
  * limitations under the License.
  */
 
-package mmarquee.automation.controls.menu;
+package mmarquee.automation.controls;
 
 import mmarquee.automation.AutomationException;
 import mmarquee.automation.ControlType;
 import mmarquee.automation.Element;
 import mmarquee.automation.ItemNotFoundException;
-import mmarquee.automation.controls.AutomationBase;
-import mmarquee.automation.controls.ElementBuilder;
 import mmarquee.automation.pattern.PatternNotFoundException;
 import mmarquee.uiautomation.TreeScope;
