+++ src/main/java/mmarquee/automation/controls/TitleBar.java
@@ -19,7 +19,6 @@ import com.sun.jna.ptr.PointerByReference;
 import mmarquee.automation.AutomationException;
 import mmarquee.automation.ControlType;
 import mmarquee.automation.Element;
-import mmarquee.automation.controls.menu.MainMenu;
 import mmarquee.uiautomation.TreeScope;
 
 /**
