--- src/main/java/mmarquee/automation/controls/menu/package-info.java
@@ -1,8 +1,0 @@
-/**
- * Domain classes used to encapsulate menu using the ms ui automation libraries.
- *
- * @since 0.4
- * @author Mark Humphreys
- * @version 0.4
- */
-package mmarquee.automation.controls.menu;
